+++ decidim-proposals/app/helpers/decidim/proposals/proposal_wizard_helper.rb
@@ -132,6 +132,19 @@ module Decidim
         end
       end
 
+      # Renders the back link except for step_2: compare
+      def proposal_wizard_aside_link_to_back(step)
+        case step
+        when :step_1
+          url = proposals_path
+        when :step_3
+          url = compare_proposal_path
+        when :step_4
+          url = edit_draft_proposal_path
+        end
+        url
+      end
+
       def wizard_aside_back_text
         case type_of
         when :collaborative_drafts